requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.6.0
    # via django
attrs==22.2.0
    # via pytest
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
coverage==6.5.0
    # via
    #   -r requirements.in
    #   coveralls
coveralls==3.3.1
    # via -r requirements.in
cryptography==41.0.1
    # via jwcrypto
deprecated==1.2.14
    # via jwcrypto
django==4.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-oauth-toolkit
    #   djangorestframework
    #   model-bakery
django-cors-headers==4.0.0
    # via -r requirements.in
django-oauth-toolkit==2.3.0
    # via -r requirements.in
django-rest-framework==0.1.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via django-rest-framework
docopt==0.6.2
    # via coveralls
exceptiongroup==1.1.1
    # via pytest
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
jwcrypto==1.5.0
    # via django-oauth-toolkit
model-bakery==1.12.0
    # via -r requirements.in
oauthlib==3.2.2
    # via django-oauth-toolkit
packaging==23.0
    # via pytest
pluggy==1.0.0
    # via pytest
psycopg2==2.9.6
    # via -r requirements.in
pycparser==2.21
    # via cffi
pytest==7.2.2
    # via pytest-django
pytest-django==4.5.2
    # via -r requirements.in
pytz==2023.3
    # via djangorestframework
requests==2.28.2
    # via
    #   coveralls
    #   django-oauth-toolkit
sqlparse==0.4.3
    # via django
tomli==2.0.1
    # via pytest
urllib3==1.26.15
    # via requests
wrapt==1.15.0
    # via deprecated
